requirements.txt
python-slugify==4.0.1
requests==2.25.1
